+++ degu/base.py
@@ -30,6 +30,8 @@ try:
         _MAX_HEADER_COUNT,
         EmptyPreambleError,
         _read_preamble,
+        format_request_preamble,
+        format_response_preamble,
     )
 except ImportError:
     from ._basepy import (
@@ -37,6 +39,8 @@ except ImportError:
         _MAX_HEADER_COUNT,
         EmptyPreambleError,
         _read_preamble,
+        format_request_preamble,
+        format_response_preamble,
     )
 
 
@@ -45,6 +49,8 @@ __all__ = (
     '_MAX_HEADER_COUNT',
     'EmptyPreambleError',
     '_read_preamble',
+    'format_request_preamble',
+    'format_response_preamble',
 )